% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longGraphs.R
\docType{package}
\name{longGraphs}
\alias{longGraphs}
\title{longGraphs: Longitudinal Graphs}
\description{
longGraphs is a package to make elegant longitudinal graphics as easy as they
should be. The entire package was built using ggplot2 functions.
}
