% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti_plot.R
\name{spaghetti_plot}
\alias{spaghetti_plot}
\title{Tidy Spaghetti Plot.}
\usage{
spaghetti_plot(
  dataf,
  id,
  x,
  y,
  group = NULL,
  wrap = NULL,
  mfrow = NULL,
  funcs = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_title = NULL,
  legend_labels = NULL
)
}
\arguments{
\item{dataf}{A dataframe or tibble.}

\item{id}{Variable with the experimental unit index.}

\item{x}{Variable to \code{x} axis.}

\item{y}{Variable to \code{y} axis (response variable).}

\item{group}{Grouping variable (if exists).}

\item{wrap}{Formula to separate plot (facet).}

\item{mfrow}{Number of rows and columns of facets.}

\item{funcs}{Functions to be added to plot.}

\item{xlab}{String with the label of \code{x} axis.}

\item{ylab}{String with the label of \code{y} axis.}

\item{legend_title}{String with the legend title.}

\item{legend_labels}{String with the legend labels.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object with the plot.
}
\description{
Function to make a Spaghetti Plot using Grammar of Graphics.
}
\examples{
spaghetti_plot(PotRoy, Individual, Age, Distance, Sex, xlab = "Age (years)",
               ylab = "Distance (cm)", legend_title = "Sex",
               legend_labels = c("Female", "Male"), wrap = ~Sex,
               funcs = mean)

}
\author{
Matheus Castro
\href{mailto::mtcastro43@gmail.com}{mtcastro43@gmail.com}
(\href{https://orcid.org/0000-0001-7879-7089}{ORCID}).
}
